refactor(agents): move agentrun api handlers
